Change the logic for whether to show the confirmation step in the Chrome extension.
The Chrome extension now stores a 1 character hash of the last-used master password, and only shows the master password confirmation step when that changes.

This is much better UX, but knowledge of that 1 chracter hash would reduce the brute force effort by 64 times, hence we don’t use this approach on the web version.