Add self destruct functionality to UI
This is primarily for the web UI (rather than in Chrome extension where the UI is reset as soon as the extension loses focus).  If the user doesn’t use the page for 2 minutes then it destroys all of the sensitive data in the page (i.e. master password and generated password).  This is to mitigate the chances of someone finding your browser unattended and stealing your data.